enhancements to share concept, component file restructing
