+++ peerdea-native/screens/ShareConceptScreen.js
@@ -2,58 +2,18 @@ import React, { Component } from 'react';
 import { Button, Image, View, StyleSheet, Text, TouchableOpacity, TextInput, Alert } from 'react-native';
 // because we're using mangaged apps version of expo (and not bare version):
 import { ImagePicker, Permissions, Camera } from 'expo';
+import { ConceptDescription } from '../components/ConceptDescription'; 
+import { ConceptAuthor } from '../components/ConceptAuthor'; 
 
 
+// PICK UP HERE
+//TODO: change infrastructure of this file to make state hold multple values of what a concet is 
 
-export class ConceptAuthor extends React.Component {
-  constructor(props) {
-    super(props);
-    this.state = { text: 'Enter your name here:' };
-  }
-
-
-  // first, should check to make sure valid group name (numbers and letters only, longer than 5 characters)
-  // then, there should be some check as to whether the group name is already taken
-  // if not already taken, submit API post request to create a new group
-  // if it is taken, catch error and say group name already taken
-  render() {
-    return (
-      <TextInput
-        style={{height: 40, borderColor: 'gray', borderWidth: 1}}
-        onChangeText={(text) => this.setState({text})}
-        value={this.state.text}
-      />
-    );
-  }
-}
-
-export class ConceptDescription extends React.Component {
-  constructor(props) {
-    super(props);
-    this.state = { text: 'Describe your concept here' };
-  }
-
-
-  // first, should check to make sure valid group name (numbers and letters only, longer than 5 characters)
-  // then, there should be some check as to whether the group name is already taken
-  // if not already taken, submit API post request to create a new group
-  // if it is taken, catch error and say group name already taken
-  render() {
-    return (
-      <TextInput
-        style={{height: 70, borderColor: 'gray', borderWidth: 1}}
-        onChangeText={(text) => this.setState({text})}
-        value={this.state.text}
-      />
-    );
-  }
-}
-
-
-export default class ConceptImagePicker extends React.Component {
+export default class ShareConcept extends React.Component {
   state = {
+    author: null,
     image: null,
-    //should probably have permissions variable in state
+    story: null,
   };
 
 
@@ -66,47 +26,42 @@ export default class ConceptImagePicker extends React.Component {
 
 
   render() {
-    let { image } = this.state;
+    let { image } = this.state //why would thi.state.image here give error this.state.image.image does not exist? where is this additional .image being appended?
 
     return (
-
       <View style={{ flex: 1, alignItems: 'center', justifyContent: 'center' }}>
-
       <Text style={styles.getStartedText}>Share a concept with your group</Text>
-
-      <ConceptAuthor/>
-
-      
-        <Button
-          title="Pick an image from camera roll"
-          onPress={this._pickImage}
-        />
-        <Button
-          title="Take a picture"
-          onPress={this._takePicture}
-        />
-        {image &&
-        <Image source={{ uri: image }} style={{ width: 200, height: 200 }} />}
-
-        <ConceptDescription/>
-
-        <Button
-          onPress={() => {
-            Alert.alert('Thanks for sharing!');
-          }}
-          title="Share concept with my group"
-          color="#841584"
-          accessibilityLabel="Share concept with my group"
-        />
+      <ConceptAuthor onChangeText={(text) => this.setState({story})}/>
+      <Button
+        title="Pick an image from camera roll"
+        onPress={this._pickImage}
+      />
+      <Button
+        title="Take a picture"
+        onPress={this._takePicture}
+      />
+      {image &&
+      <Image source={{ uri: image }} style={{ width: 200, height: 200 }} />}
+      <ConceptDescription/>
+      <Button
+        onPress={() => {
+          Alert.alert('Thanks for sharing!');
+        }}
+        title="Share concept with my group"
+        color="#841584"
+        accessibilityLabel="Share concept with my group"
+      />
       </View>
     );
   }
 
+  //underscore before function name to distinguish internal methods from the lifecycle methods of react
   _pickImage = async () => {
     await this.askPermissionsAsync();
     let result = await ImagePicker.launchImageLibraryAsync({
       allowsEditing: true,
       aspect: [4, 3],
+      base64: true,
     });
 
     // probably need some express api post call to add "result" variable to database
@@ -114,6 +69,10 @@ export default class ConceptImagePicker extends React.Component {
     //uncomment below if you'd like to see what RESULT is, printed in console in expo
     //console.log('RESULT IS' + result);
     //console.log('RESULT.URI IS' + result.uri);
+
+    console.log('RESULT IS' + result);
+    console.log('RESULT.URI IS' + result.uri);
+    console.log('RESULT.base64 IS' + result.base64);
 
     if (!result.cancelled) {
       this.setState({ image: result.uri });
@@ -125,8 +84,12 @@ export default class ConceptImagePicker extends React.Component {
     let result = await ImagePicker.launchCameraAsync({
       allowsEditing: true,
       aspect: [4, 3],
-      base64: false,
+      base64: true,
     });
+
+    console.log('RESULT IS' + result);
+    console.log('RESULT.URI IS' + result.uri);
+    console.log('RESULT.base64 IS' + result.base64);
     
     if (!result.cancelled) {
       this.setState({ image: result.uri });
